Move ResolvedDependency out of DependencyResolver
