+++ src/main/kotlin/net/minestom/dependencies/DependencyResolver.kt
@@ -17,19 +17,3 @@ interface DependencyResolver {
     @Throws(UnresolvedDependencyException::class)
     fun resolve(id: String, targetFolder: Path): ResolvedDependency
 }
-
-/**
- * Resolved Dependency.
- * Holds its coordinates (group, artifact, version), which are allowed to be empty if needed
- *
- * The contentsLocation URL represents the location of the dependency, on local storage.
- */
-data class ResolvedDependency(
-    val group: String, val name: String, val version: String,
-    val contentsLocation: URL, val subdependencies: List<ResolvedDependency>
-) {
-    fun printTree(indent: String = "") {
-        println("$indent- $group:$name:$version ($contentsLocation)")
-        subdependencies.forEach { it.printTree("$indent  ") }
-    }
-}
